Adjust NNCF to numpy 2.0 api (#2772)
### Changes

- https://numpy.org/devdocs/numpy_2_0_migration_guide.html
- Enable numpy rules for ruff

### Related tickets

138867

+++ nncf/tensor/functions/torch_numeric.py
@@ -245,7 +245,9 @@ def quantile(
             axis,
             keepdims,
         )
-    return torch.tensor(np.quantile(a.detach().cpu().numpy(), q=q, axis=axis, keepdims=keepdims)).to(device)
+    return torch.tensor(
+        np.quantile(a.detach().cpu().numpy().astype(np.float64, copy=False), q=q, axis=axis, keepdims=keepdims)
+    ).to(device)
 
 
 @numeric.percentile.register(torch.Tensor)
